🗺️ Reading Object and calling ChunkData reading
