+++ src/main/kotlin/cc/lixou/stracciatella/instance/data/IcedChunkData.kt
@@ -1,29 +1,21 @@
 package cc.lixou.stracciatella.instance.data
 
-import java.io.DataOutputStream
+import java.io.DataInputStream
 
-data class IcedChunkData(var empty: Boolean, var sections: Array<IcedSectionData>) {
+class IcedChunkData {
 
-    fun write(dos: DataOutputStream) {
-        sections.forEach { it.write(dos) }
+    companion object {
+
+        fun load(dis: DataInputStream): IcedChunkData {
+
+
+            return IcedChunkData()
+        }
+
     }
 
-    override fun equals(other: Any?): Boolean {
-        if (this === other) return true
-        if (javaClass != other?.javaClass) return false
+    fun save() {
 
-        other as IcedChunkData
-
-        if (empty != other.empty) return false
-        if (!sections.contentEquals(other.sections)) return false
-
-        return true
-    }
-
-    override fun hashCode(): Int {
-        var result = empty.hashCode()
-        result = 31 * result + sections.contentHashCode()
-        return result
     }
 
 }
